requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.25
paypalrestsdk==1.13.1
unidecode==1.3.6
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
geopy==2.4.0
pytz==2023.3
requests==2.31.0
